ramips: add support for ZyXEL LTE3302
The ZyXEL LTE3302 is an 4G indoor CPE with 2 external LTE antennas.
There are multiple variants of the device. This patch includes support for the "Deutsche Telekom" as well as the
"genuine ZyXEL" variant*.
The device is fully functional including the LTE modem.

Specifications:

    SoC: MediaTek MT7620N
    RAM: 64 MB
    Flash: 16 MB MB NAND (W25Q128FV)
    WiFi: MediaTek MT7620N built in wifi
    Switch: 2 LAN ports (100M)
    LTE: WNC M18QW Modem
    SIM: 1 micro-SIM slot
    Buttons: Reset, WPS
    LEDs: Multicolor LTE signal, multicolor internet, Wifi, LAN, Battery

The device is built as an indoor ethernet to LTE bridge or router with
Wifi.

UART Serial:

57600 8N1
Located on unpopulated 4 pin header J8:

[o] GND
[o] RX
[o] TX
[o] 3.3V Vcc

MAC assignment:
lan: 60:31:97:84:5F:EC (base, on the device back)
wlan: 60:31:97:84:5F:ED

Installation:

    Boot bootloader into recovery by pushing and holding the reset key while powering on the device
    Wait until lights start flashing quickly
    Open http://192.168.123.254
    Upload OpenWrt openwrt-ramips-mt7620-zyxel_lte3302-squashfs-factory.bin
    If http doesn't work, it can be done with curl command:
    curl -F FN=@XXXXX.bin http://192.168.123.254/upg
    where XXXXX.bin is name of firmware file.
    Wait for OpenWrt to boot and ssh to root@192.168.1.1

variants differ by their ROM_ID. You can check the ROM_ID in the bootloader using the info command.

The ROM_ID is needed for the factory firmware so that the booloader recognizes and accepts the firmware

Signed-off-by: U M <ulv@on2.de>