ramips: Zyxel LTE3302: load mac addresses and eeprom via nvmem.
Note: the device uses JBoot and as far as I understand the utility jboot_config_read, the addresses in config partition could be subject to change. Since I only own a single device, I have no way of checking if the config partition addresses change between the devices.

Note2: the config partition does only contain a single "base" mac address which is used for the LAN port. For Wifi this mac address needs to be incremented by one. There seems to be no way of doing this increment directly in the dts, which is why i added the device to the userspace "10_fix_wifi_mac" hotplug script.